+++ bagua/script/bagua_sys_perf
@@ -5,6 +5,7 @@ import torch.nn.functional as F
 import torch.optim as optim
 import torch.utils.data.distributed
 from torchvision import models
+from bagua.torch_api.algorithms.gradient_allreduce import GradientAllReduceAlgorithm
 
 # import horovod.torch as hvd
 import timeit
@@ -88,7 +89,7 @@ if args.cuda:
 optimizer = optim.SGD(model.parameters(), lr=0.01 * lr_scaler)
 
 # 用Bagua模型、优化器覆盖原始模型和优化器
-model, optimizer = bagua.bagua_init(model, optimizer)
+mode, = model.with_bagua([optimizer], GradientAllReduceAlgorithm())
 
 # # Horovod: (optional) compression algorithm.
 # compression = hvd.Compression.fp16 if args.fp16_allreduce else hvd.Compression.none
