--- bagua/torch_api/algorithms/allreduce/__init__.py
@@ -1,2 +1,0 @@
-from .allreducer import Allreducer
-from .sg_allreducer import ScatterGatherAllreducer
