--- bagua/torch_api/algorithms/decentralize/__init__.py
@@ -1,5 +1,0 @@
-"""
-Synchronized Decentralized Parallel SGD: https://arxiv.org/pdf/1705.09056.pdf
-"""
-
-from .decentralized_reducer import DecentralizedReducer
